change things

+++ app/io.h
@@ -33,6 +33,6 @@ uint64_t get_file_size(const char* filepath);
 uint64_t get_free_space(const char* device);
 int read_first_filename(char* path, char* output, size_t out_size);
 
-int get_files_in_folder(char* folder, char* out_filenames, size_t* total_folders, SearchFilter* filter, size_t max_files);
+int get_files_in_folder(char* folder, char* out_filenames, int* total_folders, SearchFilter* filter, size_t max_files);
 
 #endif
